LPS-54970 remove static methods from SassToCssBuilder
