+++ portal-impl/src/com/liferay/portal/tools/sass/SassFile.java
@@ -18,10 +18,7 @@ import com.liferay.portal.kernel.util.CharPool;
 import com.liferay.portal.kernel.util.FileUtil;
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringPool;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.servlet.filters.dynamiccss.RTLCSSUtil;
-import com.liferay.portal.tools.SassToCssBuilder;
 import com.liferay.portal.util.AggregateUtil;
 import com.liferay.portal.util.CSSBuilderUtil;
 
@@ -29,7 +26,6 @@ import java.io.File;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.Callable;
 
 /**
  * @author Minhchau Dang
@@ -51,114 +47,16 @@ public class SassFile implements SassFragment {
 		}
 	}
 
-	public void build() throws Exception {
-		long start = System.currentTimeMillis();
+	public void addSassFragment(SassFragment sassFragment) {
+		_sassFragments.add(sassFragment);
+	}
 
-		File file = new File(_docrootDirName, _fileName);
+	public String getBaseDir() {
+		return _baseDir;
+	}
 
-		if (!file.exists()) {
-			return;
-		}
-
-		String content = FileUtil.read(file);
-
-		int pos = 0;
-
-		StringBundler sb = new StringBundler();
-
-		while (true) {
-			int commentX = content.indexOf(_CSS_COMMENT_BEGIN, pos);
-			int commentY = content.indexOf(
-				_CSS_COMMENT_END, commentX + _CSS_COMMENT_BEGIN.length());
-
-			int importX = content.indexOf(_CSS_IMPORT_BEGIN, pos);
-			int importY = content.indexOf(
-				_CSS_IMPORT_END, importX + _CSS_IMPORT_BEGIN.length());
-
-			if ((importX == -1) || (importY == -1)) {
-				sb.append(content.substring(pos));
-
-				break;
-			}
-			else if ((commentX != -1) && (commentY != -1) &&
-					 (commentX < importX) && (commentY > importX)) {
-
-				commentY += _CSS_COMMENT_END.length();
-
-				sb.append(content.substring(pos, commentY));
-
-				pos = commentY;
-			}
-			else {
-				sb.append(content.substring(pos, importX));
-
-				String mediaQuery = StringPool.BLANK;
-
-				int mediaQueryImportX = content.indexOf(
-					CharPool.CLOSE_PARENTHESIS,
-					importX + _CSS_IMPORT_BEGIN.length());
-				int mediaQueryImportY = content.indexOf(
-					CharPool.SEMICOLON, importX + _CSS_IMPORT_BEGIN.length());
-
-				String importFileName = null;
-
-				if (importY != mediaQueryImportX) {
-					mediaQuery = content.substring(
-						mediaQueryImportX + 1, mediaQueryImportY);
-
-					importFileName = content.substring(
-						importX + _CSS_IMPORT_BEGIN.length(),
-						mediaQueryImportX);
-				}
-				else {
-					importFileName = content.substring(
-						importX + _CSS_IMPORT_BEGIN.length(), importY);
-				}
-
-				if (!importFileName.isEmpty()) {
-					if (importFileName.charAt(0) != CharPool.SLASH) {
-						importFileName = _fixRelativePath(
-							_baseDir.concat(importFileName));
-					}
-
-					SassFile importSassFile = SassToCssBuilder.execute(
-						_docrootDirName, importFileName);
-
-					if (Validator.isNotNull(mediaQuery)) {
-						_sassFragments.add(
-							new SassFileWithMediaQuery(
-								importSassFile, mediaQuery));
-					}
-					else {
-						_sassFragments.add(importSassFile);
-					}
-				}
-
-				// LEP-7540
-
-				if (Validator.isNotNull(mediaQuery)) {
-					pos = mediaQueryImportY + 1;
-				}
-				else {
-					pos = importY + _CSS_IMPORT_END.length();
-				}
-			}
-		}
-
-		_addSassString(_fileName, sb.toString());
-
-		String rtlCustomFileName = CSSBuilderUtil.getRtlCustomFileName(
-			_fileName);
-
-		File rtlCustomFile = new File(_docrootDirName, rtlCustomFileName);
-
-		if (rtlCustomFile.exists()) {
-			_addSassString(rtlCustomFileName, FileUtil.read(rtlCustomFile));
-		}
-
-		_elapsedTime = System.currentTimeMillis() - start;
-
-		return;
+	public String getFileName() {
+		return _fileName;
 	}
 
 	@Override
@@ -217,6 +115,10 @@ public class SassFile implements SassFragment {
 		return _rtlContent;
 	}
 
+	public void setElapsedTime(long elapsedTime) {
+		_elapsedTime = elapsedTime;
+	}
+
 	@Override
 	public String toString() {
 		StringBundler sb = new StringBundler(5);
@@ -256,50 +158,7 @@ public class SassFile implements SassFragment {
 		rtlCacheFile.setLastModified(ltrFile.lastModified());
 	}
 
-	private void _addSassString(String fileName, String sassContent)
-		throws Exception {
-
-		sassContent = sassContent.trim();
-
-		if (sassContent.isEmpty()) {
-			return;
-		}
-
-		_sassFragments.add(new SassString(fileName, sassContent));
-	}
-
-	private String _fixRelativePath(String fileName) {
-		String[] paths = StringUtil.split(fileName, CharPool.SLASH);
-
-		StringBundler sb = new StringBundler(paths.length * 2);
-
-		for (String path : paths) {
-			if (path.isEmpty() || path.equals(StringPool.PERIOD)) {
-				continue;
-			}
-
-			if (path.equals(StringPool.DOUBLE_PERIOD) && (sb.length() >= 2)) {
-				sb.setIndex(sb.index() - 2);
-
-				continue;
-			}
-
-			sb.append(StringPool.SLASH);
-			sb.append(path);
-		}
-
-		return sb.toString();
-	}
-
 	private static final String _BASE_URL = "@base_url@";
-
-	private static final String _CSS_COMMENT_BEGIN = "/*";
-
-	private static final String _CSS_COMMENT_END = "*/";
-
-	private static final String _CSS_IMPORT_BEGIN = "@import url(";
-
-	private static final String _CSS_IMPORT_END = ");";
 
 	private final String _baseDir;
 	private final String _docrootDirName;
